Student/AttendanceControllerにQueryServiceの適用（共通化）
